refactor: centralize site url lookup
