+++ app/digital-garden/page.tsx
@@ -1,6 +1,6 @@
 import Link from 'next/link'
 import type { Metadata } from 'next'
-import { cookies, headers } from 'next/headers'
+import { cookies } from 'next/headers'
 import { getAllNotes } from '@/lib/digital-garden'
 import { getSiteName } from '@/lib/settings'
 import DigitalGardenList from '@/components/digital-garden-list'
@@ -8,6 +8,7 @@ import { Badge } from '@/components/ui/badge'
 import { cn } from '@/lib/utils'
 import en from '@/locales/en.json'
 import es from '@/locales/es.json'
+import { getCanonicalUrl } from '@/utils/getCanonicalUrl'
 
 const translations = { en, es } as const
 
@@ -23,11 +24,7 @@ export async function generateMetadata(): Promise<Metadata> {
   const cookieStore = cookies()
   const locale = (cookieStore.get('NEXT_LOCALE')?.value || 'en') as 'en' | 'es'
   const t = getT(locale)
-  const headersList = headers()
-  const host =
-    headersList.get('x-forwarded-host') || headersList.get('host') || 'localhost:3000'
-  const protocol = headersList.get('x-forwarded-proto') || 'https'
-  const siteUrl = process.env.NEXT_PUBLIC_SITE_URL || `${protocol}://${host}`
+  const siteUrl = getCanonicalUrl()
   const url = locale === 'es' ? `${siteUrl}/es/digital-garden` : `${siteUrl}/digital-garden`
   const title = `${siteName}'s ${t('navbar.garden')}`
   return {
