+++ app/robots.ts
@@ -1,7 +1,8 @@
 import type { MetadataRoute } from "next"
+import { getCanonicalUrl } from "@/utils/getCanonicalUrl"
 
 export default function robots(): MetadataRoute.Robots {
-  const siteUrl = process.env.NEXT_PUBLIC_SITE_URL || "https://fabri.lat"
+  const siteUrl = getCanonicalUrl()
   return {
     rules: {
       userAgent: "*",
